+++ mini_packages.json
@@ -17,7 +17,7 @@
     "@react-navigation/native": "6.1.3",
     "@react-navigation/stack": "6.3.12",
     "@shopify/flash-list": "1.4.1",
-    "@sleeperhq/mini-core": "1.9.4",
+    "@sleeperhq/mini-core": "1.9.5",
     "amazon-cognito-identity-js": "6.3.2",
     "crypto-js": "3.3.0",
     "decimal.js-light": "2.5.1",
